[go] Relocate move generating functions to a new module
