+++ src/go/kataPlay.ts
@@ -1,6 +1,7 @@
 import type { GoOpponent, NS } from 'netscript';
 
 import { GtpClient, toIndices, toVertex } from 'go/GtpClient';
+import { randomNearInvalidMove, getRandomMove } from 'go/moves';
 import { filterMapBoard, Move, Node, Vertex } from 'go/types';
 
 import { CONFIG } from 'go/config';
@@ -105,50 +106,4 @@ async function playGame(ns: NS, client: GtpClient) {
             }
         }
     }
-}
-
-function randomNearInvalidMove(
-    validMoves: boolean[][],
-    [x, y]: [number, number],
-): [number, number] {
-    const moveOptions = [];
-    for (let i = x - 2; i <= x + 2; i++) {
-        if (i < 0 || i >= validMoves.length) continue;
-        for (let j = y - 2; j <= y + 2; j++) {
-            if (j < 0 || j >= validMoves[i].length) continue;
-            if (validMoves[i][j]) {
-                moveOptions.push([i, j]);
-            }
-        }
-    }
-    // Choose one of the found moves at random
-    const randomIndex = Math.floor(Math.random() * moveOptions.length);
-    return moveOptions[randomIndex] ?? [-1, -1];
-}
-
-/**
- * Choose one of the empty points on the board at random to play
- */
-function getRandomMove(board: string[], validMoves: boolean[][]) {
-    const moveOptions = [];
-    const size = board[0].length;
-
-    // Look through all the points on the board
-    for (let x = 0; x < size; x++) {
-        for (let y = 0; y < size; y++) {
-            // Make sure the point is a valid move
-            const isValidMove = validMoves[x][y] === true;
-            // Leave some spaces to make it harder to capture our pieces.
-            // We don't want to run out of empty node connections!
-            const isNotReservedSpace = x % 2 === 1 || y % 2 === 1;
-
-            if (isValidMove && isNotReservedSpace) {
-                moveOptions.push([x, y]);
-            }
-        }
-    }
-
-    // Choose one of the found moves at random
-    const randomIndex = Math.floor(Math.random() * moveOptions.length);
-    return moveOptions[randomIndex] ?? [-1, -1];
 }
